custom_user_model/requirements.txt
django-verfied-email-field
django-extensions
